refactor: 퇴원차트에서 입원차트 이동 버튼 수정
