+++ components/hospital/icu/main/out-and-visit/out-and-visit-skeleton.tsx
@@ -2,5 +2,5 @@ import { Skeleton } from '@/components/ui/skeleton'
 import React from 'react'
 
 export default function OutAndVisitSkeleton() {
-  return <Skeleton className="h-[400px] w-full" />
+  return <Skeleton className="h-[320px] w-full" />
 }
